+++ functional/build.gradle
@@ -1,0 +1,16 @@
+plugins {
+    id 'org.jetbrains.kotlin.jvm'
+    id 'org.jmailen.kotlinter'
+    id 'java-test-fixtures'
+}
+
+repositories {
+    jcenter()
+}
+
+dependencies {
+    testFixturesApi gradleTestKit()
+    testFixturesApi 'junit:junit:4.13'
+    testFixturesApi 'com.nhaarman.mockitokotlin2:mockito-kotlin:2.2.0'
+    testFixturesApi 'org.jetbrains:annotations:18.0.0'
+}
