+++ functional/kotlin/src/test/kotlin/ExtensionTest.kt
@@ -1,17 +1,16 @@
-package org.jmailen.gradle.kotlinter.functional
-
+import functional.WithGradleTest
+import functional.utils.kotlinClass
+import functional.utils.resolve
+import functional.utils.settingsFile
 import groovy.util.GroovyTestCase.assertEquals
 import java.io.File
 import org.gradle.testkit.runner.TaskOutcome
 import org.intellij.lang.annotations.Language
-import org.jmailen.gradle.kotlinter.functional.utils.kotlinClass
-import org.jmailen.gradle.kotlinter.functional.utils.resolve
-import org.jmailen.gradle.kotlinter.functional.utils.settingsFile
 import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Test
 
-internal class ExtensionTest : WithGradleTest.Kotlin() {
+internal class ExtensionTest : WithGradleTest() {
 
     lateinit var projectRoot: File
