+++ functional/src/testFixtures/kotlin/functional/utils/Factories.kt
@@ -1,21 +1,21 @@
-package org.jmailen.gradle.kotlinter.functional.utils
+package functional.utils
 
 import org.intellij.lang.annotations.Language
 
 @Language("kotlin")
-internal fun kotlinClass(className: String) = """
+fun kotlinClass(className: String) = """
     object $className
     
 """.trimIndent()
 
 @Language("groovy")
-internal val settingsFile = """
+val settingsFile = """
     rootProject.name = 'kotlinter'
     
 """.trimIndent()
 
 @Language("xml")
-internal val androidManifest = """
+val androidManifest = """
      <manifest package="com.example" />
 
 """.trimIndent()
