+++ functional/src/testFixtures/kotlin/functional/utils/Utilities.kt
@@ -1,0 +1,9 @@
+package functional.utils
+
+import java.io.File
+
+fun File.resolve(path: String, receiver: File.() -> Unit): File =
+    resolve(path).apply {
+        parentFile.mkdirs()
+        receiver()
+    }
