+++ settings.gradle.kts
@@ -1,1 +1,3 @@
 rootProject.name = "kotlinter-gradle"
+
+include(":functional:android", ":functional:kotlin")
